server-appengine V3/requirements.txt
Flask==2.1.0
Pandas==1.3.5
google-cloud-ndb
scipy
pyspark
pyparsing
joblib
scikit-learn